[MIRROR] [MDB IGNORE] Improves Tramstation xenobio pens (#4723)
Original PR: https://github.com/tgstation/tgstation/pull/73661
-----
## About The Pull Request
Adds another set of windoors to the Tramstation Xenobiology pens,
opposite from the ones already present, and refactors those windoors to
use `base_state = "right"; icon_state = "right";` instead of `dir`.
## Why It's Good For The Game
This removes one space from the pens, but ensures that everything inside
a pen is visible from the floor above them, so there won't be any hidden
monkeys or slimes sitting in the square that the windoors are in. If
this square is meant to be used as part of the pen, then the lower floor
should instead be made more accessible to the slime console, having
fewer walls and more windows so all slimes can be seen at once. This
also adds pens that are one space large for working on one slime at a
time, though these tiny pens are not usable from the floor above. This
change also makes the tramstation xenobio layout more like how it is on
other maps. This also makes it more obvious how the windoors are setup,
as the map file shows that they are explicitly set to close to the
right, rather than changing their dir to an abitrary number.
## Changelog
:cl:
qol: Added another set of windoors to the Tramstation Xenobiology pens,
fully enclosing the space before them. The creatures in the main pens
can now always be accessed from above.
refactor: Tramstation Xenobiology pens explicitly close to the right
rather than being rotated by using dir.
/:cl:

---------

Co-authored-by: Shroopy <delyth@shroop.net>