[MIRROR] [MDB IGNORE] stylistic alterations of galoshes icons (#4748)
Original PR: https://github.com/tgstation/tgstation/pull/73688
-----
## About The Pull Request
updates the galoshes icons to make them stylistically similar to the
work boots and jackboots, with the palate of the insulated gloves. also
similarly modified the absorbent galoshes, but used color palate of the
purple hat.
this includes the item's inventory icon, the worn icon and the in hand
icons.

## Why It's Good For The Game

it is much more aesthetically pleasing than the flat shaded, neon curry
colored ones. also, it makes me feel pretty when I wear them with
insulated gloves on.

![image](https://user-images.githubusercontent.com/121331795/221738653-86ebc317-cf59-4524-ba8a-dcdf8528710e.png)



## Changelog

:cl:

imageadd: added prettier icons for the normal and absorbent galoshes

/:cl:

---------

Co-authored-by: Chlorotrifluoride <121331795+Chlorotrifluoride@users.noreply.github.com>