[MIRROR] [MDB IGNORE] Updates Coderbus and Adminbus Ruin (#4764)
Original PR: https://github.com/tgstation/tgstation/pull/73675
-----

## About The Pull Request

Updates the people inside the coderbus and adminbus. Removes the
meatballs, because... well, they used to be called something that
actually accurately describes the admin team but now they're just
meatballs. The banhammer now takes the driver seat of adminbus, with a
sacrificial altar in the back. RIP to all the good players lost when
they became admin

## Why It's Good For The Game

It was really out of date lol

## Changelog
:cl:
add: Updated the bus ruin
/:cl:

---------

Co-authored-by: tralezab <40974010+tralezab@users.noreply.github.com>