Security Aesthetics Improvement Finale (LAP 2) (#73795)
## About The Pull Request
No parts to this, no bullshit, straight to the point.

This PR finishes respriting things that weren't tackled by either me or
Viro recently, with the exception of the riot and bulletproof gear.


![image](https://user-images.githubusercontent.com/77556824/222966086-ba5515ec-8f58-49e4-ac6c-7ae7e00a4c86.png)

Head of Security outfits received a general upgrade (including a global
redo of the beret to fit better shading standards)


![image](https://user-images.githubusercontent.com/77556824/222966120-19cd65c9-5190-4cfb-9560-fc1e439e7a63.png)

The same goes for warden clothing (Obj sprite now properly reflects the
appearance of the mob sprite)


![image](https://user-images.githubusercontent.com/77556824/222966159-6ca5bc51-3822-4326-8066-3d815de84bb3.png)


![image](https://user-images.githubusercontent.com/77556824/222966170-e784f49d-c16d-49d5-b480-acc8966b5020.png)

Security officer clothing (including jackboots and vest) have also
received upgrades, with either minor reshades, or complete overhauls to
their styling.


![image](https://user-images.githubusercontent.com/77556824/222966490-cd996952-5ec1-4819-a2f9-6fe359782790.png)

Aerostatic Suit has been edited into a shirt, to better reflect the
reference character's default outfit, which is a plain white shirt, and
not a button-up. The vest has also been resprited into a flak vest!
## Why It's Good For The Game
They're definite improvements three months in the making. I hope to god
that this time, they actually last, in this never-ending race to out-do
myself and the competition!
## Changelog
:cl: PositiveEntropy
imageadd: Tweaked, Overhauled and Resprited a variety of garments
related to the security department, particularly its alternate clothes!
Check it when possible!
/:cl: