[MIRROR] [MDB IGNORE] [NO GBP] fixes typo in PR template (#4837)
Original PR: https://github.com/tgstation/tgstation/pull/73928
-----
## About The Pull Request

There is a verb in the English language that has the function of
expressing the meaning of getting something from someone or something
else that has given it to you or sent it to you or delivered it to you,
being given something by someone or something else that has decided to
give it to you or send it to you or deliver it to you, or being the
recipient of something that someone or something else has given to you
or sent to you or delivered to you. This verb is “receive.” This word
has a very specific spelling that follows a very common rule in the
English language that states “i before e except after c.” However, some
people may have difficulty remembering this rule or applying this rule
correctly when they write this word. They may write this word with an
“-ie-” combination instead of an “-ei-” one. This is always a very big
mistake and should be avoided in your writing at all costs. The word
“recieve” does not exist in the English language and will be marked as
an error by any spell checker or grammar tool that you use.

## Why It's Good For The Game

I would like to express my sincere and heartfelt apology for spelling
the word receive wrong in my previous message. I made a very big mistake
and wrote the word with an “-ie-” combination instead of an “-ei-” one.
This is always incorrect and does not follow the very common rule in the
English language that states “i before e except after c.” The word
receive has a very specific spelling that follows this rule. The word
recieve does not exist in the English language and will be marked as an
error by any spell checker or grammar tool that I use. I am very sorry
for this mistake and I hope that you will forgive me for it.

## Changelog

I would like to inform you that there is no changelog available for this
update, because this update does not constitute a change to the game
itself. A changelog is a document that records the changes that have
been made to the game in each update. However, this update does not make
any changes to the game itself. It only fixes some minor issues that
were affecting the performance of the game. Therefore, there is no need
for a changelog for this update, because there are no changes to the
game itself that need to be recorded in a document.

Co-authored-by: tralezab <40974010+tralezab@users.noreply.github.com>