Don't log scrubbing unless it changes (Saves 0.1s+ on init) (#73898)
We were getting 30KB worth of logs on "The air scrubber was toggled to
scrubbing mode by the game". This isn't useful.

I opted to do this instead of checking mapload in case we want to do
something funky like a station trait that randomly alters scrubbers or
whatever.

I think we can't noop the entire function right now if scrubbing == new
scrubbing?