Resprites Donut Boxes (#73896)
## About The Pull Request
Resprites the donut boxes to look better. and to be more accurate since
the resprites of the actual donuts.

![image](https://user-images.githubusercontent.com/75904284/224395088-973270be-0664-41f0-8eb7-83990d5f8354.png)

![image](https://user-images.githubusercontent.com/75904284/224493041-3f347cb1-9ed7-4a25-83b2-a7791d1b9913.png)
Old sprite for reference

![image](https://user-images.githubusercontent.com/75904284/224493157-661ca228-8e96-45d2-aece-677f409b4e33.png)

![image](https://user-images.githubusercontent.com/75904284/224493176-f7f3ce75-564b-4f6b-972c-f18daacb9558.png)

## Why It's Good For The Game
I think it looks better, and it removes some very old sprites. Also
shifts the sprite to be in 3/4ths.
## Changelog
-Changes all sprites for the donut box, and relating to the donut box.
:cl:
imageadd: Resprites Donut Box, and all sprites for donuts in it.
/:cl: