Cleans up wall mounts and Issues (#4849)
Jolly asked that I take a look at the issue tracker and fix some issues.
I've also taken the liberty to de-obfuscate as many wall-mounts in
engineering as I could. Most things are on their own walls where
logically possible. All APC's should be on their own walls now, since
this is the only thing that actually mechanically matters. If the rest
of the station is like this, then I can understand why an issue was made
for it. It's bad.
Also cleans out the following Issues:
closes #4726
closes #4727

partially addresses  #4150
I wouldn't go calling it a culling, but actually spreading things out to
a reasonable degree. I moved some request consoles away from desks with
this, because those are mounts that I can actually shove in corners.


## Changelog
:cl:
fix: A cleanup pass on Limastation mostly concentrated around
engineering - For now.
/:cl: