[MIRROR] [MDB IGNORE] Optimizes atmos init by 20% (#4854)
Original PR: https://github.com/tgstation/tgstation/pull/73926
-----

## About The Pull Request

Each time we intialized a turf's atmos, we checked all the turfs around
it to see if they were different. This meant each pair of turfs talked
to each other twice.

If we instead do the comparing in a second loop, we can use
current_cycle to ensure we only compare a pair once. This saves 0.5
seconds of atmos init.

## Why It's Good For The Game

Speed

---------

Co-authored-by: LemonInTheDark <58055496+LemonInTheDark@users.noreply.github.com>