Fixes the issues with ghost role spawners being able to spawn userless bodies (#73636)
## About The Pull Request
Turns out there was no verification to ensure that you weren't already
in the process of trying to join a ghost role, when clicking on a ghost
role spawner.

That resulted in people being able to double-click on a spawner and
basically, trying to spawn twice... Or thrice... Or even more than that
if they're fast.

Of course, that's not a good thing, so I made it so you can only try to
spawn one time at a time, which includes only getting one prompt at a
time. That includes ALL spawners, because I know that bug will otherwise
pop up at some point from someone thinking they're clever.

I added a CRASH() for the handling of ckey-less ghosts, in case someone
tries to do another subtype that somehow manages to bypass all of the
checks I already added. That way, they'll know if they fuck up.

I also made a second uses check considering there's stoplags in the
is_banned() check.

Fixes https://github.com/tgstation/tgstation/issues/73619.

## Why It's Good For The Game
Having mindless bodies is just not very cash money and it can break
spawners to the point of making them go to negative amounts of uses,
which isn't very cash money.

## Changelog

:cl: GoldenAlpharex
fix: You should only be able to spawn as one ghost role at a time. Close
the prompt if you want to spawn as another one. Begone soulless
randomgen humans!
/:cl: