[MIRROR] [MDB IGNORE] Fixes some areas on IceBoxStation (#4876)
Original PR: https://github.com/tgstation/tgstation/pull/73910
-----

## About The Pull Request

Basically these still followed the old pattern (pre-#69975) where you
would have the area also be outside of the room in order to power the
lights. i just went through and patched those since I saw them today and
it's too trivial to nag someone else to do it
### Mapping March
(does this count for that? i'm out of the loop and just noticed that
something was outdated when someone asked me to double-check a map)
Ckey to recieve rewards: san7890

## Why It's Good For The Game

this is the mapping pattern now, and i sure as hell don't want to go
back or otherwise insinuate that things should still be done this way
## Changelog
this doesn't matter as far as players are concerned

---------

Co-authored-by: san7890 <the@san7890.com>