[MIRROR] [MDB IGNORE] Cyborg surgery processor shows downloaded surgeries & equip requirement on examine (#4914)
Original PR: https://github.com/tgstation/tgstation/pull/73992
-----
## About The Pull Request
Surgery processor can download surgeries to be accessed without a
surgical computer. However, it does not tell the player that it needs to
be actually in one of the three equipped modules to let the cyborg
actually use those surgeries without relying on a surgical computer.

This PR addresses that + makes it show downloaded surgeries.


![image](https://user-images.githubusercontent.com/75863639/225000403-59cbf370-1c50-44a1-8758-2dab98876802.png)

## Why It's Good For The Game
Closes #71120
Players should be able to use the processor without any complications or
misinterpretations of its intended function. Being able to check
downloaded surgeries for something specific on the go can be handy.

## Changelog
:cl:
qol: medical cyborg's surgical processor now tells the user on examine
that it needs to be equipped in an active module to access downloaded
surgeries
qol: medical cyborg's surgical processor now lists downloaded surgeries
on examine
/:cl:

---------

Co-authored-by: Sealed101 <75863639+Sealed101@users.noreply.github.com>