[MIRROR] [MDB IGNORE] Fixes inconsistent list syntax in `set_status_tab` (#4983)
Original PR: https://github.com/tgstation/tgstation/pull/74079
-----
BYOND is inconsistent about whether it treats these as `"key" = value`,
or `src.key = value`, or something else. We suspect that this behavior
is fluctuating in 515, and given that this proc has been causing immense
overtime, that it might be the culprit of some recent complains of
recurring lag. And if it's not, then we at least have more consistent
code.

---------

Co-authored-by: Mothblocks <35135081+Mothblocks@users.noreply.github.com>