[MIRROR] [MDB IGNORE] Adds the Janitor access keyring (#5002)
Original PR: https://github.com/tgstation/tgstation/pull/73768
-----
## About The Pull Request

HackMD for this can be found here:
https://hackmd.io/VEbjO1kaQJarao4KqGfzgw?view

Basically, this gives the first Janitor an access key that they can use
to enter departments. This requires a Head of Staff to approve it
through the keycard authenticator, and only holds one department at a
time, and will clear itself out after 10 minutes.
This gives departmental AA, including the head of staff's office, adding
a tradeoff.

The Janitor is alerted when access is given to their key (as their
keyring will say such) and everyone can see janitors trying to open a
door when they use it (with a cool sound, too).
You can't bump open airlocks with it, I limited it to directly using it
on the door, because I thought it made it feel more realistic to a
keyring.

I had an earlier version of this that gave the key it's own access,
which allowed bumping open, but it also allowed things like locking RD
consoles, and I dunno how strong I want this key to be.

I also wanted to limit how easy this is to greytide/steal, currently
only the first Janitor spawns with the key. I thought it would be too
easy to exploit otherwise, or essentially stolen roundstart if there
were no Janitors. Maybe this can also fit as a Traitor objective after
melon's new objective PR is done?

### Minor detail

While adding icons, I realized inhands don't actually set its icon state
to the default if it's null, so I removed that LYING comment.

## Why It's Good For The Game


![geromekey](https://user-images.githubusercontent.com/53777086/222824948-4d42d801-fea5-423b-9050-2a23d02f8ea4.gif)

![gerome](https://user-images.githubusercontent.com/53777086/222824598-1225aeb0-4aa8-4826-bf29-5674e9cda7ad.gif)

Alternatively, https://youtu.be/dlkSbQ-IkRM?t=182 (timestamp)

## Changelog

:cl: JohnFulpWillard, sprites by BalkyGoat
add: The Janitor Access key: Janitors can now be given departmental AA
(one at a time) using the Keycard authentication device in Command
offices. Use it if you need cleaning!
fix: HoP's trim is no longer set to edit Supply access.
/:cl:

---------

Co-authored-by: John Willard <53777086+JohnFulpWillard@users.noreply.github.com>
Co-authored-by: Fikou <23585223+Fikou@users.noreply.github.com>
Co-authored-by: tattle <66640614+dragomagol@users.noreply.github.com>
Co-authored-by: Zephyr <12817816+ZephyrTFA@users.noreply.github.com>