Changes the Heads of Staffs equipment slightly (CMO, HoS, RD, CE) (#4978)
Going forward with the lore team, some changes will be happening with
the chain of command and things that apply to them directly.
Since we haven't moved forward yet with discussions, I'm implementing
changes I can make in the meantime.

While doing this I've realized the CE, HoP and Captain lack significant
drip. This is a tragedy.

For the unknown, they all have turtlenecks now.

![image](https://user-images.githubusercontent.com/70232195/226101785-1b3c4e4d-ca24-44cc-b9c4-b1d2d59b6c5f.png)

## Changelog


:cl: Jolly 
add: In preperation for eventual job-lore changes, the CE, RD, CMO and
HoS had their outfits tweaked ever so slightly. This is purely cosmetic.
Turtlenecks ahoy!
/:cl: