[MIRROR] [MDB IGNORE] Fix missing meat clown food types (#5014)
Original PR: https://github.com/tgstation/tgstation/pull/74146
-----
:cl: coiax
fix: Fixed missing food types for the meat clown, it is now meat and
fruit.
/:cl:

Liked by lizards, toxic to moths, and neutral to humans and felnids in
terms of species biases.

The construction recipe is steak + banana, but the resulting product has
no food types, making it blandly suitable for everyone. This is an
oversight. Not everyone will like a mixture of banana and flesh.

---------

Co-authored-by: Jack Edge <yellowbounder@gmail.com>