[MIRROR] [MDB IGNORE] The Lance Crew Evacuation System (New Emergency Shuttle) (#5019)
Original PR: https://github.com/tgstation/tgstation/pull/74024
-----
## About The Pull Request
Yeah this shuttle deliberately hits the station.
While more hazardous to people in departures, the Lance is a great
crew-friendly tool for a station that has a severely damaged departures!
No longer will you need to struggle with firelocks and collapse just
outside the shuttle in a depressurized departures, when The Lance comes
RIGHT TO YOU.

<details>
<summary> The Lance being used </summary>


![image](https://cdn.discordapp.com/attachments/585862469508005888/1087777600535003266/image.png)

</details>

<details>

<summary> Map screenshot! </summary>


![image](https://user-images.githubusercontent.com/73589390/226687950-3844ade3-ee90-4056-b923-c1a9b460eb61.png)

</details>

### Mapping March
Ckey to receive rewards: Cheshify 
I don't need a token or anything, but the GBP or whatever would be nice
so I can make more map stuff

## Why It's Good For The Game

New shuttles are good! This one has both immense pros (very safe when
you're inside it, lots of supplies) and immense cons (smashes into the
fucking station). It'll be an interesting purchase for the antagonist
using it to kill a bunch of people in departures, and a useful tool for
the crew making it easier for them to evacuate.

This shuttle is not restricted to emags or any other prerequisite, as
that would defeat the purpose. It's supposed to be a mixture of pros and
cons, and keeping it locked behind an emag removes it's use as a
prohibitively expensive but ultimately useful tool for the crew.
## Changelog
:cl:
add: Nanotrasen has given the crew access to a brand new emergency
shuttle! Watch out for it's destructive entrance!
/:cl:

---------

Co-authored-by: Cheshify <73589390+Cheshify@users.noreply.github.com>